requirements.txt
SpeechRecognition
pyttsx3
fuzzywuzzy
python-dotenv
openai
